requirements.txt
django>=5.0,<7
djangorestframework>=3.15
psycopg2-binary>=2.9
python-dotenv>=1.0
django-cors-headers>=4.0
reportlab>=4.0
